+++ src/main/java/org/sosly/villageworks/VillageWorks.java
@@ -1,8 +1,18 @@
 package org.sosly.villageworks;
 
 import com.mojang.logging.LogUtils;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.RegisterCommandsEvent;
+import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.slf4j.Logger;
+import org.sosly.villageworks.command.VillageWorksCommand;
+import org.sosly.villageworks.entity.Villager;
+import org.sosly.villageworks.registry.EntityTypes;
 
 @Mod(VillageWorks.MOD_ID)
 public class VillageWorks {
@@ -11,5 +21,28 @@ public class VillageWorks {
 
     public VillageWorks() {
         LOGGER.info("Loading VillageWorks");
+
+        IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
+
+        EntityTypes.register(modEventBus);
+
+        modEventBus.addListener(this::setup);
+        modEventBus.addListener(this::onEntityAttributeCreation);
+
+        MinecraftForge.EVENT_BUS.register(this);
+    }
+
+    private void setup(final FMLCommonSetupEvent event) {
+        LOGGER.info("VillageWorks setup complete");
+    }
+
+    @SubscribeEvent
+    public void onEntityAttributeCreation(EntityAttributeCreationEvent event) {
+        event.put(EntityTypes.VILLAGER.get(), Villager.createAttributes().build());
+    }
+
+    @SubscribeEvent
+    public void onRegisterCommands(RegisterCommandsEvent event) {
+        VillageWorksCommand.register(event.getDispatcher());
     }
 }