+++ src/main/java/org/sosly/villageworks/registry/EntityTypes.java
@@ -1,0 +1,26 @@
+package org.sosly.villageworks.registry;
+
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.MobCategory;
+import net.minecraftforge.common.ForgeSpawnEggItem;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
+import org.sosly.villageworks.VillageWorks;
+import org.sosly.villageworks.entity.Villager;
+
+public class EntityTypes {
+    public static final DeferredRegister<EntityType<?>> ENTITY_TYPES = 
+        DeferredRegister.create(ForgeRegistries.ENTITY_TYPES, VillageWorks.MOD_ID);
+
+    public static final RegistryObject<EntityType<Villager>> VILLAGER = 
+        ENTITY_TYPES.register("villager", () -> EntityType.Builder.of(Villager::new, MobCategory.CREATURE)
+            .sized(0.6F, 1.95F)
+            .clientTrackingRange(10)
+            .build("villager"));
+
+    public static void register(IEventBus eventBus) {
+        ENTITY_TYPES.register(eventBus);
+    }
+}