+++ CHANGELOG.md
@@ -5,4 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/SoSly/VillagerLiteracyProgram/tree/1.20.1)
+## [Unreleased](https://github.com/SoSly/VillageWorks/tree/1.20.1)
+
+### Added
+- A new `Villager` entity that is the basis of the mod
+- `/vw assign` command to set villager homes